pyproject.toml
[project]
name = "valid_length"
version = "0.0.0"
description = "Validates that the length of value is within the expected range."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "rstr",
    "guardrails-ai>=0.3.2",
    "pydantic>=2.4.2"
]